Update with actual CalendarEventsGroupTest.java
